Refactor feature extractor classes
Split classes in the `featureextractor` module up into seperate files.
Previously, all were in featureextractor.py, but now split into:

- __init__.py: imports all other classes *directly* now

- applauseextractor.py: `ApplauseExtractor` class for spectral
centroid-based applause detection

- constqsegmentextractor.py: `ConstQSegmentExtractor` class for
segmentation

- enums.py: all feature extractor-based enums, currently `BlockDomain`
only.

- febase.py: `GenericExtractor` class

- loudnessextractor.py: `LoudnessExtractor` class, for RMS-based
loudness extraction.
