+++ cleverthumbnailer/featureextractor/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+__author__ = 'Jon'
+"""Audio feature extraction classes for use with BBC audio thumbnail chooser application"""
+
+from enums import BlockDomain
+from febase import GenericExtractor
+from loudnessextractor import LoudnessExtractor
+from constqsegmentextractor import ConstQSegmentExtractor
+from applauseextractor import ApplauseExtractor
